requirements.txt
# Core dependencies for AI HR Assistant
openai>=1.12.0
langchain>=0.2.0,<0.3.0
langchain-openai>=0.1.0,<0.2.0
langchain-community>=0.2.0,<0.3.0
gradio>=4.15.0
chromadb>=0.4.22
pypdf>=4.0.0
python-dotenv>=1.0.0

# Text processing and utilities
tiktoken>=0.5.2
numpy>=1.24.0,<2.0.0
pandas>=2.0.0

# Optional: For local embeddings (if not using OpenAI)
sentence-transformers>=2.2.0

# Development and testing
pytest>=7.4.0
black>=23.0.0